worker/requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.1
kombu==5.2.3
packaging==21.3
prompt-toolkit==3.0.28
psycopg-binary==3.0.9
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
